refactor: theme colors and fonts and fix like icon fill color
