+++ src/components/Icons/ui/Heart.Icon.tsx
@@ -39,11 +39,13 @@ const Heart = ({ width, fill = '' }: HeartProps) => {
           viewBox="467 392 58 57"
           xmlns="http://www.w3.org/2000/svg"
         >
-          <g fill="none" fillRule="evenodd" transform="translate(467 392)">
+          <g fill="none" fillRule="nonzero" transform="translate(467 392)">
             <path
               d="M29.144 20.773c-.063-.13-4.227-8.67-11.44-2.59C7.63 28.795 28.94 43.256 29.143 43.394c.204-.138 21.513-14.6 11.44-25.213-7.214-6.08-11.377 2.46-11.44 2.59z"
               id="heart"
-              fill={clicked ? fill : '#666'}
+              fill={clicked ? fill : 'transparent'}
+              stroke={clicked ? fill : 'gray'}
+              strokeWidth="1"
             />
             <circle
               id="main-circ"
