+++ src/styles/global.ts
@@ -1,5 +1,5 @@
 import { css } from '@emotion/core';
-import { fonts } from '../gatsby-plugin-theme-ui';
+import { fonts } from '../gatsby-plugin-theme-ui/fonts';
 
 export const globalStyles = theme => css`
   /**
